kernel: kernel+modules build for ARM as well as amd64 now
The cross-compiler patches have been updated accordingly, and a wide range of issues related to pointer alignment have also been cleaned up.

+++ src/modules/subsys/posix/PosixSyscallManager.h
@@ -50,7 +50,7 @@ class PosixSyscallManager : public SyscallHandler
      *\note Not implemented (singleton) */
     PosixSyscallManager &operator=(const PosixSyscallManager &);
     /** Records seen unknown syscalls so we don't spam logs. */
-    Tree<size_t, bool> m_SeenUnknownSyscalls;
+    Tree<uint64_t, bool> m_SeenUnknownSyscalls;
 };
 
 #endif
